+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 
-### Next
+### 2020.4
+
+- Support 4 finger swipe gestures.
+- Fix that smart zoom settings are not persisted.
+
+### Open Source
 
 - Open source.
 - Support to build without license management logics.